+++ actionview/test/template/log_subscriber_test.rb
@@ -11,10 +11,13 @@ class AVLogSubscriberTest < ActiveSupport::TestCase
   def setup
     super
 
-    view_paths     = ActionController::Base.view_paths
+    view_paths = ActionController::Base.view_paths
+    view_paths.each(&:clear_cache)
+    ActionView::LookupContext.fallbacks.each(&:clear_cache)
+
     lookup_context = ActionView::LookupContext.new(view_paths, {}, ["test"])
     renderer       = ActionView::Renderer.new(lookup_context)
-    @view          = ActionView::Base.new(renderer, {})
+    @view          = ActionView::Base.with_empty_template_cache.new(renderer, {})
 
     ActionView::LogSubscriber.attach_to :action_view